CHECK POINT <FINISH 100%> ...

+++ bonus/server_bonus.c
@@ -6,7 +6,7 @@
 /*   By: abadouab <abadouab@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/24 11:34:17 by abadouab          #+#    #+#             */
-/*   Updated: 2024/04/17 17:25:45 by abadouab         ###   ########.fr       */
+/*   Updated: 2024/04/18 18:18:31 by abadouab         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,38 +21,56 @@ static void	error_handler(void)
 	exit (EXIT_FAILURE);
 }
 
-void	check_pid(int *old_pid, int new_pid, int *mess, int *bits)
+static void	check_pid(int new_pid, t_data *data)
 {
-	if (*old_pid != new_pid)
+	if (data->old_pid != new_pid)
 	{
-		*bits = 0;
-		*mess = 0;
+		data->bits = 0;
+		data->set = 0;
+		data->bytes = 0;
+		data->mess = 0;
 		g_set_bit = 128;
 	}
-	*old_pid = new_pid;
+	data->old_pid = new_pid;
+}
+
+static void	message_handler(t_data *data)
+{
+	if (data->mess >= 128)
+		data->bytes = 4;
+	else if (data->mess >= 224)
+		data->bytes = 3;
+	else if (data->mess >= 192)
+		data->bytes = 2;
+	if (data->bytes != data->set)
+		data->message[data->set++] = data->mess;
+	if (!data->bytes)
+		ft_printf("%c", data->mess);
+	else if (data->bytes == data->set)
+	{
+		ft_printf("%s", data->message);
+		data->bytes = 0;
+		data->set = 0;
+	}
+	data->bits= 0;
+	data->mess = 0;
+	g_set_bit = 128;
 }
 
 static void	signal_handler(int signal_client, siginfo_t *sig_inf, void *none)
 {
-	static int	old_pid = 0;
-	static int	bits = 0;
-	static int	mess = 0;
+	static t_data	data;
 
 	(void)none;
-	check_pid(&old_pid, sig_inf->si_pid, &mess, &bits);
+	check_pid(sig_inf->si_pid, &data);
 	if (signal_client == SIGUSR1)
-		mess |= g_set_bit;
+		data.mess |= g_set_bit;
 	g_set_bit >>= 1;
-	bits++;
-	if (bits == 8)
-	{
-		if (!mess)
-			kill(sig_inf->si_pid, SIGUSR1);
-		ft_printf("%c", mess);
-		bits = 0;
-		mess = 0;
-		g_set_bit = 128;
-	}
+	data.bits++;
+	if (data.bits == 8 && data.mess)
+		message_handler(&data);
+	if (data.bits == 8 && !data.mess)
+		kill(sig_inf->si_pid, SIGUSR1);
 }
 
 int	main(int ac, char **av)
